Rename ImageStreamTagReconciler to PromotionReconciler
The old name was oriented towards the type the controller works on, not
the problem it solves. This becomes confusing once there is more than
one controller that operates on the given type.

This PR renames the controller to indicate its purpose.
