+++ pkg/controller/promotionreconciler/testdata/imagestreamtag.yaml
