+++ pkg/controller/promotionreconciler/testdata/ist_with_git_suffix.yaml
